Remove separate binding storage, store values in TulispObjects directly
This makes the VM slower and takes away the possibility to have
lexical scoping, but that probably needs a different approach anyway -
one with garbage collection or some other way of cleanup.

But for now, this allows for variable bindings to be shared between
the tree walking interpreter and the virtual machine.